+++ test/parallel/test-repl-console.js
@@ -1,13 +1,10 @@
 'use strict';
 var common = require('../common'),
     assert = require('assert'),
-    Stream = require('stream'),
     repl = require('repl');
 
-// create a dummy stream that does nothing
-var stream = new Stream();
-stream.write = stream.pause = stream.resume = function() {};
-stream.readable = stream.writable = true;
+// Create a dummy stream that does nothing
+const stream = new common.ArrayStream();
 
 var r = repl.start({
   input: stream,
