+++ test/parallel/test-repl-syntax-error-stack.js
@@ -11,24 +11,12 @@ process.on('exit', () => {
   assert.strictEqual(found, true);
 });
 
-// A stream to push an array into a REPL
-function ArrayStream() {
-  this.run = function(data) {
-    data.forEach(line => {
-      this.emit('data', line + '\n');
-    });
-  };
-}
-util.inherits(ArrayStream, require('stream').Stream);
-ArrayStream.prototype.readable = true;
-ArrayStream.prototype.writable = true;
-ArrayStream.prototype.resume = function() {};
-ArrayStream.prototype.write = function(output) {
+common.ArrayStream.prototype.write = function(output) {
   if (/var foo bar;/.test(output))
     found = true;
 };
 
-const putIn = new ArrayStream();
+const putIn = new common.ArrayStream();
 const testMe = repl.start('', putIn);
 let file = path.resolve(__dirname, '../fixtures/syntax/bad_syntax');
